Moved everything to the common module, added BungeeCord support
Feature/bungee (#6)

* Begin implementing Bungeecord support, rename Catalyst to CatalystVelocity and CatalystBungee

* Moved LuckpermsUtils to Common, extracted more code from the chat listeners into the ChatService

+++ api/src/main/java/org/anvilpowered/catalyst/api/service/TabService.java
@@ -21,10 +21,9 @@ public interface TabService<TString> {
 
     TString formatTab(String format, String userName, String prefix, String suffix);
 
-    TString formatPlayerSpecificTab(String format, String userName, String prefix, String suffix, int ping, int playerCount);
+    TString formatPlayerSpecificTab(String format, String userName, String prefix, String suffix, long ping, int playerCount);
 
     String getBalance(String userName);
 
     void setBalance(String userName, double balance);
-
 }
