+++ common/src/main/java/org/anvilpowered/catalyst/common/service/CommonTabService.java
@@ -49,7 +49,7 @@ public class CommonTabService<TString, TCommandSource> implements TabService<TSt
     }
 
     @Override
-    public TString formatPlayerSpecificTab(String format, String userName, String prefix, String suffix, int ping, int playerCount) {
+    public TString formatPlayerSpecificTab(String format, String userName, String prefix, String suffix, long ping, int playerCount) {
         format = format.replace("%player%", userName)
             .replace("%prefix%", prefix)
             .replace("%suffix%", suffix)
@@ -68,7 +68,7 @@ public class CommonTabService<TString, TCommandSource> implements TabService<TSt
 
     @Override
     public void setBalance(String userName, double balance) {
-        if(containsKey(userName)) {
+        if (containsKey(userName)) {
             playerBalances.replace(userName, balance);
         } else {
             playerBalances.put(userName, balance);
